+++ app/src/pages/Dashboard/EditEventsPage/EditEventsPage.tsx
@@ -34,7 +34,7 @@ const StepButton = styled(MuiStepButton)({
 const Container = styled(Box)(({ theme }) => ({
   display: 'grid',
   gridTemplateColumns: '1fr',
-  gridTemplateRows: 'auto auto 1fr',
+  gridTemplateRows: 'auto auto auto 1fr',
   height: '100%',
   overflow: 'hidden auto',
   gap: theme.spacing(2),
@@ -112,6 +112,10 @@ const EditEventsPage = () => {
     <Container>
       <Typography component="h1">{pageHeading}</Typography>
       <StepperHeader />
+      <Typography>
+        <span role="presentation">*</span>
+        {t('dashboard-create-or-update-required-fields-info')}
+      </Typography>
       {activeStep === 0 && (
         <CreateOrUpdateMeetingForm existingEvent={event} onForwardButtonClick={() => setActiveStep(1)} />
       )}